Separeted verifyToken function to its own file
